# swagger.yaml
openapi: 3.0.0
info:
  version: "1"
  title: "Blood-Donation API"
  description: "This is a blood donation api project. anyone can register for an account and will be able to login to their credential. anyone can request for blood"
  contact:
    name: Rayhan-Abdullah
    url: https://coder-rayhan.vercel.app/
    email: programmer.rayhan1@gmail.com
servers:
  - description: Dev
    url: "http://localhost:4000/api/v1"
...
paths:
...
  "/volunteers/{id}":
    get:
      security:
        - bearerAuth: []
      description: Get a Single Volunteer
      tags: [Volunteer]
      parameters:
        - in: path
          name: id
          required: true
          description: volunteer ID
          schema:
            type: string
            example: 'sldflasdfladfadf'
      responses:
        "200":
          description: Return a Single Volunteer
          content:
            "application/jons":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  data:
                    allOf:
                      - $ref: "#/components/schemas/VolunteerData"
                      - $ref: "#/components/schemas/Timestamp"
                  links:
                    type: string
                    example: "/volunteers/asdlfalsdfasf33aslf"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
    delete:
      security:
        - bearerAuth: []
      description: Delete a Volunteer Requests
      tags: [Volunteer]
      parameters:
        - in: path
          name: id
          required: true
          description: volunteer ID
          schema:
            type: string
            example: 'sldflasdfladfadf'
      responses:
        "204":
          $ref: "#/components/responses/204"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          example: "10:33:21"
        updatedAt:
          type: string
          example: "10:33:21"
      required: [createdAt, updatedAt]
...
    VolunteerData:
      type: object
      properties:
        age:
          type: integer
          format: int32
          example: 26
        occupation:
          type: string
          example: student
        bloodGroup:
          type: string
          example: o+
        gender:
          type: string
          example: male
        phone: 
          type: string
          example: 01798716196
        study:
          type: string
          example: honours 3rd year
        cover:
          type: string
          format: url
          example: http://w3.unsplas.com
        address:
          type: object
          properties:
            division:
              type: string
              example: barishal
            dist:
              type: string
              example: Bhola
            upazila:
              type: string
              example: Charfassion
            streetAddress:
              type: string
              example: Charfassion
            currentAddress:
              type: string
              example: Dhaka
          required: [division, dist, upazilla, currentAddress, streetAddress]
        nationalId:
          type: string
          example: 0983734826324
        bio:
          type: string
          example: Hi, i am rayhan abdullah, i am programmer. i always try to help other..
      required: [_id, age, occupation, gender, bloodGroup, address]
  responses:
...
    "500":
      description: Server Error
      content:
        "application/json":
          schema:
            properties:
              code:
                type: integer
                format: int32
                example: 500
              message:
                type: string
                example: "Server Error!"
            required: [code, message]
    "401":
      description: Unauthorized
      content:
        "application/json":
          schema:
            properties:
              code:
                type: integer
                format: int32
                example: 401
              message:
                type: string
                example: "Unauthorized!"
            required: [code, message]
    "404":
      description: Not Found
      content:
        "application/json":
          schema:
            properties:
              code:
                type: integer
                format: int32
                example: 404
              message:
                type: string
                example: "Not Found"
            required: [code, message]
    "204":
      description: Deleted Successfully Done
      content:
        "application/json":
          schema:
            properties:
              code:
                type: integer
                format: int32
                example: 204
              message:
                type: string
                example: "Successfully Deleted"
            required: [code, message]
...